+++ Diffinitely/Models/PRCommentItem.cs
@@ -1,12 +1,16 @@
 using System.Collections.ObjectModel;
+using System.ComponentModel;
+using System.Runtime.CompilerServices;
 using System.Runtime.Serialization;
 using Microsoft.VisualStudio.Extensibility.UI;
 
 namespace Diffinitely.Models;
 
 [DataContract]
-public class PrCommentItem
+public class PrCommentItem : INotifyPropertyChanged
 {
+    private string _replyText = "";
+
     [DataMember]
     public long CommentId { get; set; }
 
@@ -21,6 +25,9 @@ public class PrCommentItem
 
     [DataMember]
     public DateTimeOffset CreatedAt { get; set; }
+
+    [DataMember]
+    public string FormattedCreatedAt => CreatedAt.ToString("yyyy-MM-dd HH:mm");
 
     [DataMember]
     public string Body { get; set; } = "";
@@ -38,10 +45,44 @@ public class PrCommentItem
     public bool CanResolve { get; set; }
 
     [DataMember]
+    public bool CanUnresolve { get; set; }
+
+    [DataMember]
+    public bool IsOutdated { get; set; }
+
+    [DataMember]
+    public bool CanReply { get; set; }
+
+    [DataMember]
+    public bool CanJumpToDiff { get; set; }
+
+    [DataMember]
+    public string ReplyText
+    {
+        get => _replyText;
+        set
+        {
+            if (_replyText != value)
+            {
+                _replyText = value;
+                RaisePropertyChanged();
+            }
+        }
+    }
+
+    [DataMember]
     public ObservableCollection<PrCommentReply> ThreadReplies { get; set; } = new();
 
     [DataMember] public IAsyncCommand? ResolveCommand { get; set; }
+    [DataMember] public IAsyncCommand? UnresolveCommand { get; set; }
     [DataMember] public IAsyncCommand? ViewCommand { get; set; }
+    [DataMember] public IAsyncCommand? ReplyCommand { get; set; }
+    [DataMember] public IAsyncCommand? JumpToDiffCommand { get; set; }
+
+    public event PropertyChangedEventHandler? PropertyChanged;
+
+    protected void RaisePropertyChanged([CallerMemberName] string? propName = null)
+        => PropertyChanged?.Invoke(this, new PropertyChangedEventArgs(propName));
 }
 
 [DataContract]
@@ -49,5 +90,6 @@ public class PrCommentReply
 {
     [DataMember] public string Author { get; set; } = "";
     [DataMember] public DateTimeOffset CreatedAt { get; set; }
+    [DataMember] public string FormattedCreatedAt => CreatedAt.ToString("yyyy-MM-dd HH:mm");
     [DataMember] public string Body { get; set; } = "";
 }